refactor: move types to own module
